+++ index.js
@@ -2,8 +2,13 @@ import React from 'react'
 import { render } from 'react-dom'
 
 import Demo from './src/js/components/demo'
+import { homepage, name, description } from './package.json'
 
 render(
-  <Demo/>,
+  <Demo
+    title={name}
+    description={description}
+    repositoryUrl={homepage}
+  />,
   document.getElementById('app')
 )