+++ src/js/export.js
@@ -1,3 +1,4 @@
 import Demo from './components/demo'
+import HowToStart from './components/how-to-start'
 
-export default Demo
+export { Demo, HowToStart }